Make batch job conflict with add-tested and use-merge-strategy
*  Rename the group to make it clearer these are experimental features.
*  Rename --experimental-batch to --batch. The help message marks it as
experimental instead, which is consistent with the --add-merge-strategy
option.

Note that putting the batch option in this group means the merge
strategy can't be used with batching. There doesn't seem to be a good
way of using the same argument in different mutually exclusive groups,
and this seems acceptable for now given they are both experimental
features.